refactor: fix colors, sorts, routes, cs
