+++ package.json
@@ -66,6 +66,7 @@
     "@wdio/mocha-framework": "^8.33.1",
     "@wdio/spec-reporter": "^8.32.4",
     "axios": "^1.6.7",
+    "babel-plugin-jsx-remove-data-test-id": "^3.0.0",
     "craco-alias": "^3.0.1",
     "eslint": "^8.56.0",
     "eslint-import-resolver-alias": "^1.1.2",
