+++ src/App.tsx
@@ -1,6 +1,8 @@
 import React, { useContext, useEffect, useState } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 
+import { getAuth } from 'firebase/auth';
+
 import MainRoutes from './routes/MainRoutes';
 import Header from '@components/Header/Header';
 import Login from '@components/Login/Login';
@@ -8,7 +10,6 @@ import { FirebaseContext } from '@context/FirebaseContext';
 import { ThemeProvider } from '@emotion/react';
 import { createTheme } from '@mui/material';
 import ThemeToggler from '@UI/ThemeToggler/ThemeToggler';
-import { getAuth } from 'firebase/auth';
 
 import './styles/reset.css';
 import './styles/media.css';
