+++ src/UI/SearchInput/SearchInput.module.css
@@ -42,7 +42,7 @@
   height: 12px;
   min-width: 12px;
   border-radius: 50%;
-  border: 2px solid grey;
+  border: 2px solid var(--img-background);
   position: relative;
 }
 
@@ -50,7 +50,7 @@
   content: '';
   width: 14px;
   height: 2px;
-  background-color: grey;
+  background-color: var(--img-background);
   display: block;
   position: absolute;
   border-radius: 10px;
