+++ src/UI/ThemeToggler/ThemeToggler.module.css
@@ -30,7 +30,6 @@
   width: 80px;
   height: 4px;
   border-radius: 2px;
-  background-color: black;
   position: absolute;
 }
