+++ src/components/BookItems/BooksItem/BookItem.tsx
@@ -1,12 +1,13 @@
 import React, { useContext, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 
+import { getAuth } from 'firebase/auth';
+
 import { FirebaseContext } from '@context/FirebaseContext';
 import { useTypedSelector } from '@hooks/useTypedSelector';
 import { Typography } from '@mui/material';
 import { IBook } from '@projectTypes/IBooks';
 import { addToDB, deleteFromBD, hasBook } from '@utils/Firebase';
-import { getAuth } from 'firebase/auth';
 
 import styles from './BookItem.module.css';
