+++ src/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Provider } from 'react-redux';
+import ReactDOM from 'react-dom/client';
 
 import App from './App';
 import ErrorBoundary from '@components/ErrorBoundary/ErrorBoundary';
 import { FirebaseContext } from '@context/FirebaseContext';
 import { store } from '@store/index';
 import { app, db } from '@utils/Firebase';
-import ReactDOM from 'react-dom/client';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
