+++ src/pages/FavoritesPage/FavoritesPage.tsx
@@ -1,12 +1,13 @@
 import React, { useContext, useEffect, useState } from 'react';
 
+import { getAuth } from 'firebase/auth';
+
 import BookItems from '@components/BookItems/BookItems';
 import { FirebaseContext } from '@context/FirebaseContext';
 import { useTypedSelector } from '@hooks/useTypedSelector';
 import { Typography } from '@mui/material';
 import { IBook } from '@projectTypes/IBooks';
 import { getBooks } from '@utils/Firebase';
-import { getAuth } from 'firebase/auth';
 
 function FavoritesPage() {
   const { theme } = useTypedSelector((state) => state.theme);
