--- src/routes/Routes.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import BookPage from '@pages/BookPage/BookPage';
-import FavoritesPage from '@pages/FavoritesPage/FavoritesPage';
-import MainPage from '@pages/MainPage/MainPage';
-
-export const mainPage = <MainPage />;
-export const bookPage = <BookPage />;
-export const favoritesPage = <FavoritesPage />;
