+++ src/styles/reset.css
@@ -5,5 +5,5 @@
 
 a {
   text-decoration: none;
-  color: black;
+  color: var(--font-color-dark);
 }
