+++ src/styles/vars.css
@@ -10,4 +10,6 @@
   --authors_type-light-color: #ffffff8f;
   --authors_type-dark-color: #0000007a;
   --title-light-color: rgb(221, 221, 221);
+  --img-background: #888888;
+  --heart: #ff0000;
 }
